requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
idna==2.8
plaid-python==3.4.0
pytz==2019.2
requests==2.22.0
sqlparse==0.3.0
stripe==2.36.2
urllib3==1.25.5
